(PC-17987) home: reorganise exclusivity module to separate design and offer bloc
	- ExclusivityModule content is now ExclusivityOffer
	- extract exclusivity image in a dedicated component
	- ExclusivityOffer renders a touchable link that encapsulate en ExclusivityImage
	- define a new component ExclusivityBanner that render only an ExclusivityImage
	- ExclusivityModule now returns an ExclusivityOffer when an offer ID is provided,
	  or an ExclusivityBanner if not
